+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/camera/CameraVideoCapturerFactory.kt
@@ -1,4 +1,5 @@
-// 現在は一種類のカメラキャプチャしか使用していないが、将来的に複数のカメラキャプチャーを実装する可能性があるため、インターフェースを定義している
+// 現在は一種類のカメラキャプチャしか使用していないが、将来的に複数のカメラキャプチャーを実装する可能性があるため、
+// インターフェースを定義している。
 // 将来的に、複数のカメラキャプチャーをサポートする場合は、このインターフェースを実装した複数のファクトリを用意することができる
 package jp.shiguredo.sora.sample.camera
 
@@ -12,11 +13,10 @@ interface CameraVideoCapturerFactory {
 
 class DefaultCameraVideoCapturerFactory(
     private val context: Context,
-    private val fixedResolution: Boolean = false,
     private val frontFacingFirst: Boolean = true
 ) : CameraVideoCapturerFactory {
 
     override fun createCapturer(): CameraVideoCapturer? {
-        return CameraCapturerFactory.create(context, fixedResolution, frontFacingFirst)
+        return CameraCapturerFactory.create(context, frontFacingFirst = frontFacingFirst)
     }
 }