+++ samples/src/main/kotlin/jp/shiguredo/sora/sample/ui/SimulcastActivity.kt
@@ -57,7 +57,7 @@ class SimulcastActivity : AppCompatActivity() {
     private var spotlightUnfocusRid: SoraVideoOption.SpotlightRid? = null
     private var simulcastEnabled: Boolean = true
     private var fps: Int = 30
-    private var fixedResolution = false
+    private var degradationPreference: SoraVideoOption.DegradationPreference? = null
     private var resolutionAdjustment: SoraVideoOption.ResolutionAdjustment? = null
     private var simulcastRid: SoraVideoOption.SimulcastRid? = null
     private var clientId: String? = null
@@ -165,10 +165,13 @@ class SimulcastActivity : AppCompatActivity() {
             else -> true
         }
 
-        fixedResolution = when (intent.getStringExtra("RESOLUTION_CHANGE")) {
-            "可変" -> false
-            "固定" -> true
-            else -> false
+        degradationPreference = when (intent.getStringExtra("RESOLUTION_CHANGE")) {
+            "未指定" -> null
+            "MAINTAIN_RESOLUTION" -> SoraVideoOption.DegradationPreference.MAINTAIN_RESOLUTION
+            "MAINTAIN_FRAMERATE" -> SoraVideoOption.DegradationPreference.MAINTAIN_FRAMERATE
+            "BALANCED" -> SoraVideoOption.DegradationPreference.BALANCED
+            "DISABLED" -> SoraVideoOption.DegradationPreference.DISABLED
+            else -> null
         }
 
         resolutionAdjustment = when (intent.getStringExtra("RESOLUTION_ADJUSTMENT")) {
@@ -387,7 +390,7 @@ class SimulcastActivity : AppCompatActivity() {
             simulcast = simulcastEnabled,
             simulcastRid = simulcastRid,
             videoFPS = fps,
-            fixedResolution = fixedResolution,
+            degradationPreference = degradationPreference,
             resolutionAdjustment = resolutionAdjustment,
             videoCodec = videoCodec,
             videoBitRate = videoBitRate,